[onert] Re-organize onert test directory
The directory structure in the test dir is outdated. This commit makes
it to have the same structure with src dir.

ONE-DCO-1.0-Signed-off-by: Hanjoung Lee <hanjoung.lee@samsung.com>
